+++ src/constants.ts
@@ -1,0 +1,7 @@
+export const PLACEHOLDER = "{SCID}";
+export const METHOD = "tdw";
+export const PROTOCOL = `did:${METHOD}:1`;
+export const BASE_CONTEXT = [
+  "https://www.w3.org/ns/did/v1",
+  "https://w3id.org/security/multikey/v1"
+];
