+++ src/routes/did.ts
@@ -3,7 +3,7 @@ import { resolveDID } from '../method';
 export const getLatestDIDDoc = async ({params: {id}, set}: {params: {id: string;}; set: any;}) => {
   console.log(`Resolving ${id}...`);
   try {
-    const didLog = await Bun.file(`./out/${id}/did.jsonl`).text();
+    const didLog = await Bun.file(`./test/logs/${id}/did.jsonl`).text();
     // console.log(didLog)
     // const logLine: string = '[{"op":"replace","path":"/proof/proofValue","value":"z128ss1..."}]';
     const logEntries: DIDLog = didLog.trim().split('\n').map(l => JSON.parse(l));
