+++ src/utils.ts
@@ -1,7 +1,83 @@
+import base32 from 'base32';
+import { canonicalize } from 'json-canonicalize';
+import { createHash } from 'node:crypto';
+import { nanoid } from 'nanoid';
+
 export const clone = (input: any) => JSON.parse(JSON.stringify(input));
 
 export const createDate = (created?: Date) => new Date(created ?? Date.now()).toISOString().slice(0,-5)+'Z';
 
 export function bytesToHex(bytes: Uint8Array): string {
   return Array.from(bytes).map(byte => byte.toString(16).padStart(2, '0')).join('');
+}
+
+export const createSCID = async (logEntryHash: string): Promise<string> => {
+  return `${logEntryHash.slice(0, 28)}`;
+}
+
+export const deriveHash = async (input: any): Promise<{logEntryHash: string}> => {
+  const data = canonicalize(input);
+  const hash = createHash('sha256').update(data).digest();
+  return {logEntryHash: base32.encode(hash)};
+}
+
+export const createDIDDoc = async (options: CreateDIDInterface): Promise<{doc: DIDDoc}> => {
+  const {controller} = options;
+  const {all} = normalizeVMs(options.verificationMethods, controller);
+  return {
+    doc: {
+      "@context": [
+        "https://www.w3.org/ns/did/v1",
+        "https://w3id.org/security/multikey/v1"
+      ],
+      id: controller,
+      controller,
+      ...all
+    }
+  };
+}
+
+export const createVMID = (vm: VerificationMethod, did: string | null) => {
+  return `${did ?? ''}#${vm.publicKeyMultibase?.slice(-8) || nanoid(8)}`
+}
+
+export const normalizeVMs = (verificationMethod: VerificationMethod[] | undefined, did: string | null = null) => {
+  if (!verificationMethod) {
+    return {};
+  }
+  const all: any = {};
+  const authentication = verificationMethod
+    ?.filter(vm => vm.type === 'authentication').map(vm => createVMID(vm, did))
+  if (authentication && authentication?.length > 0) {
+    all.authentication = authentication;
+  }
+  const assertionMethod = verificationMethod
+    ?.filter(vm => vm.type === 'assertionMethod').map(vm => createVMID(vm, did))
+  if (assertionMethod && assertionMethod?.length > 0) {
+    all.assertionMethod = assertionMethod;
+  }
+  const keyAgreement = verificationMethod
+    ?.filter(vm => vm.type === 'keyAgreement').map(vm => createVMID(vm, did));
+  if (keyAgreement && keyAgreement?.length > 0) {
+    all.keyAgreement = keyAgreement;
+  }
+  const capabilityDelegation = verificationMethod
+    ?.filter(vm => vm.type === 'capabilityDelegation').map(vm => createVMID(vm, did));
+  if (capabilityDelegation && capabilityDelegation?.length > 0) {
+    all.capabilityDelegation = capabilityDelegation;
+  }
+  const capabilityInvocation = verificationMethod
+  ?.filter(vm => vm.type === 'capabilityInvocation').map(vm => createVMID(vm, did));
+  if (capabilityInvocation && capabilityInvocation?.length > 0) {
+    all.capabilityInvocation = capabilityInvocation;
+  }
+  if(verificationMethod && verificationMethod.length > 0) {
+    all.verificationMethod = verificationMethod?.map(vm => ({
+      id: createVMID(vm, did),
+      ...(did ? {controller: vm.controller ?? did} : {}),
+      type: 'Multikey',
+      publicKeyMultibase: vm.publicKeyMultibase
+    }))
+  }
+  return {all};
 }
