+++ test/fixtures/not-authorized.log
@@ -1,2 +1,2 @@
-["7cqgcne0whyw7ggfb6u3kjmtjwv7x4jyau69p4g2yu76phpxrf60",1,"2024-04-16T14:27:39Z",{"method":"did:tdw:1","scid":"e9q26uqq5hg20yydmzq63fkcx7af"},{"value":{"@context":["https://www.w3.org/ns/did/v1","https://w3id.org/security/multikey/v1"],"id":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af","controller":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af","authentication":["did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#5b48Zj6B"],"assertionMethod":["did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CpsUrTh3"],"verificationMethod":[{"id":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#5b48Zj6B","controller":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af","type":"Multikey","publicKeyMultibase":"z6Mkw1KSvGWNAwSwWbcpwPgFARX4vKPa1xvcDMsJ5b48Zj6B"},{"id":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CpsUrTh3","controller":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af","type":"Multikey","publicKeyMultibase":"z6MkuocstfAaHsJgRnQgfQdtJiECuWEMDAbQV61aCpsUrTh3"}]}},[{"type":"DataIntegrityProof","cryptosuite":"eddsa-jcs-2022","verificationMethod":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#5b48Zj6B","created":"2024-04-16T14:27:39Z","proofPurpose":"authentication","challenge":"7cqgcne0whyw7ggfb6u3kjmtjwv7x4jyau69p4g2yu76phpxrf60","proofValue":"zk6RJKxo7yniuzg6aXQdanDXQPo7aZQtdotE9dQrjNLMex7yRihUqNoNzt9tHD1w2LJCAwHsrzUG4zLtyPdKvKEn"}]]
-["1k84mahn3an1u9p87hhd0ugywmxa6m44p9hfhfp2fchnt9vd7dz0",2,"2024-04-16T14:27:39Z",{},{"patch":[{"op":"replace","path":"/verificationMethod/1/publicKeyMultibase","value":"z6Mkq7Qoa2LCvLCUiq9W2J9vXH1ooDqSX2ehWGUzCZSeAWLE"},{"op":"replace","path":"/verificationMethod/1/id","value":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CZSeAWLE"},{"op":"replace","path":"/verificationMethod/0/publicKeyMultibase","value":"z6MksNvDtgMutKMzjezRqj8JbYuCHNjHR5HBaF2ogJqvqaYD"},{"op":"replace","path":"/verificationMethod/0/id","value":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#gJqvqaYD"},{"op":"replace","path":"/assertionMethod/0","value":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CZSeAWLE"},{"op":"remove","path":"/authentication"},{"op":"replace","path":"/controller","value":["did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af"]}]},[{"type":"DataIntegrityProof","cryptosuite":"eddsa-jcs-2022","verificationMethod":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CpsUrTh3","created":"2024-04-16T14:27:39Z","proofPurpose":"authentication","challenge":"1k84mahn3an1u9p87hhd0ugywmxa6m44p9hfhfp2fchnt9vd7dz0","proofValue":"z3umcqZoSFUYLS9zWF2VekrngTydcfXgQbja7WmuXfb7MnFKPkTxfciwajzkxtWH4Sd5Mep7A6hc1V7c9GbjM98FU"}]]
+["657efywtfv57hc39t0nc33j2c9n47q1enfgukm6qxc8zkpzdy9z0",1,"2024-05-24T05:57:35Z",{"method":"did:tdw:1","scid":"e9q26uqq5hg20yydmzq63fkcx7af"},{"value":{"@context":["https://www.w3.org/ns/did/v1","https://w3id.org/security/multikey/v1"],"id":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af","controller":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af","authentication":["did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#5b48Zj6B"],"assertionMethod":["did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CpsUrTh3"],"verificationMethod":[{"id":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#5b48Zj6B","controller":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af","type":"Multikey","publicKeyMultibase":"z6Mkw1KSvGWNAwSwWbcpwPgFARX4vKPa1xvcDMsJ5b48Zj6B"},{"id":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CpsUrTh3","controller":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af","type":"Multikey","publicKeyMultibase":"z6MkuocstfAaHsJgRnQgfQdtJiECuWEMDAbQV61aCpsUrTh3"}]}},[{"type":"DataIntegrityProof","cryptosuite":"eddsa-jcs-2022","verificationMethod":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#5b48Zj6B","created":"2024-05-24T05:57:35Z","proofPurpose":"authentication","challenge":"657efywtfv57hc39t0nc33j2c9n47q1enfgukm6qxc8zkpzdy9z0","proofValue":"zwKc7oAfSwJ7qbCNLdQo8Uyik5ErfpSgr21mNFRKtd8tfwYHHDNaGLJdTJstx7ZMRtmk8VeJ5dgrfFb4HZU7g7Sr"}]]
+["uezevg9kav7xme31f57yczb6mkrbdwnhyjumyk8xgvzb1n0rg3h0",2,"2024-05-24T05:57:35Z",{},{"patch":[{"op":"replace","path":"/verificationMethod/1/publicKeyMultibase","value":"z6Mkq7Qoa2LCvLCUiq9W2J9vXH1ooDqSX2ehWGUzCZSeAWLE"},{"op":"replace","path":"/verificationMethod/1/id","value":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CZSeAWLE"},{"op":"replace","path":"/verificationMethod/0/publicKeyMultibase","value":"z6MksNvDtgMutKMzjezRqj8JbYuCHNjHR5HBaF2ogJqvqaYD"},{"op":"replace","path":"/verificationMethod/0/id","value":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#gJqvqaYD"},{"op":"replace","path":"/assertionMethod/0","value":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CZSeAWLE"},{"op":"replace","path":"/authentication/0","value":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#gJqvqaYD"},{"op":"replace","path":"/controller","value":["did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af"]}]},[{"type":"DataIntegrityProof","cryptosuite":"eddsa-jcs-2022","verificationMethod":"did:tdw:example.com:e9q26uqq5hg20yydmzq63fkcx7af#CpsUrTh3","created":"2024-05-24T05:57:35Z","proofPurpose":"authentication","challenge":"uezevg9kav7xme31f57yczb6mkrbdwnhyjumyk8xgvzb1n0rg3h0","proofValue":"z43Q7RFveZjemYWvnncv4h6wsYrkXBLSjYYSpF1wbot2anj2KBX82eDeSvd2qk17vyd9y4yj4sFLvbcYQrhTwLwiA"}]]
